fix(DockerGPU): fixing docker gpu using wrong python version. Trying to fix dependency issus (WIP)
There is an issue that the tensorflow installation doesn't find the libnvinfer files. This does not however seem needed for tensorflow to work with the GPU. The files are existing  and the environemnt variable is set correctly but it just isn't picked up. Other changes totthis update include the fix of the incorrect python update to the dockerGPU (which was done incorrectly previously) , Utils have alse been created for scripts to be placed that need to run on the start of the devcontainer.